platform/polyglot/app_yaml_endpoints/py/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "app-yaml-endpoints"
version = "1.1.0"
description = "Configuration-driven endpoint routing SDK"
license = {text = "MIT"}
requires-python = ">=3.11,<3.12"
keywords = ["fetch", "router", "config", "endpoint", "sdk"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pyyaml==6.0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.5.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app_yaml_endpoints"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
